==> 04-Individual-Populations/Bootstrapping-400-subpops/03_Create_RDSs_for_NCD01s.Rmd <==
```{r echo=FALSE,  message=FALSE, warning=FALSE}
# Load required packages
libraries <- c("tidyverse", "data.table")
for (lib in libraries) {
  if (require(package = lib, character.only = TRUE)) {
    print("Successful")
  } else {
    print("Installing")
    if (!requireNamespace("BiocManager", quietly = TRUE)) {
      install.packages("BiocManager")
    }
    BiocManager::install(lib)
    library(lib, character.only = TRUE)
  }
}

```

```{r echo=FALSE,  message=FALSE, warning=FALSE}
# Loop over each directory (set-1 to set-100)
for (set_i in 1:100) {
  # Loop over each file within the directory (pop1_inputNCD.txt to pop4_inputNCD.txt)
  for (pop_i in 1:4) {
    
    input_file_path <- paste0("set-", set_i, "/pop", pop_i, "_inputNCD.txt")
    output_file_path <- paste0("set-", set_i, "/pop", pop_i, "_input_NCD.rds")
    
    # Load input
    all_chrs <- read_table(input_file_path, 
                           col_names = c("CHR","POS", "ID","REF","ALT","AF2"))
    
    all_chrs$AF1 <- (1 - all_chrs$AF2)
    all_chrs$MAF <- pmin(all_chrs$AF1,all_chrs$AF2)
    
    all_chrs_ncd_format <- data.table(CHR = all_chrs$CHR, POS = all_chrs$POS,
                                      ID = all_chrs$ID, REF = all_chrs$REF, 
                                      ALT = all_chrs$ALT, AF1 = all_chrs$AF1,
                                      AF2 = all_chrs$AF2, MAF = all_chrs$MAF)
    
    # Loop through each chromosome
    vect_chr <- unique(all_chrs$CHR)
    ncd_input_list <- list()
    
    for (chr in vect_chr) {
      chr_ncd_format     <- all_chrs_ncd_format[all_chrs_ncd_format$CHR == chr,]
      ncd_input_list[[chr]] <- chr_ncd_format
    }
    
    saveRDS(ncd_input_list, file = output_file_path)
  }
}

```
